little bit of clean up
